// src/proto-ts/common/common.pb.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               v5.29.3
// source: common/common.proto

/* eslint-disable */

export const protobufPackage = "grpcBoilerplate.common";

export interface Empty {
}

/**
 * Response Metadata
 * This message schema is for response metadata
 */
export interface ResponseMetadata {
  /**
   * Status
   * This field is used for status of the response
   */
  status: string;
  /**
   * Code
   * This field is used for code of the response
   */
  code: string;
  /**
   * Message
   * This field is used for message of the response
   */
  message: string;
}

/**
 * Global Response
 * This message schema is used for global response
 */
export interface GlobalResponse {
  /**
   * Metadata
   * This field is used for metadata of the response
   */
  metadata: ResponseMetadata | undefined;
}

/**
 * Pagination Request
 * This message schema is used for pagination request
 */
export interface PaginationRequest {
  /**
   * Page
   * This field is used for page of the pagination request
   */
  page: number;
  /**
   * Items Per Page
   * This field is used for items per page of the pagination request
   */
  itemsPerPage: number;
}

/**
 * Pagination Metadata
 * This message schema is used for pagination response metadata
 */
export interface PaginationMetadata {
  /**
   * Page Items
   * This field is used for page items of the pagination response metadata
   */
  pageItems: number;
  /**
   * Current Page
   * This field is used for current page of the pagination response metadata
   */
  currentPage: number;
  /**
   * Total Pages
   * This field is used for total pages of the pagination response metadata
   */
  totalPages: number;
  /**
   * Total Items
   * This field is used for total items of the pagination response metadata
   */
  totalItems: number;
}

export const GRPC_BOILERPLATE_COMMON_PACKAGE_NAME = "grpcBoilerplate.common";